refactor(security-auditor): extract scan and report modules (#677)
Split src/security-auditor.ts (992 -> 176-line facade) by responsibility:
- security-auditor-scan.ts: SCAN_PATTERNS, scanCode, analyzeSource,
  analyzePermissions
- security-auditor-report.ts: formatSecurityReport(+JSON) and all
  formatting helpers
Facade keeps calculateVerdict + auditSkillSecurity orchestration and
re-exports the full prior surface; sibling describes move to
security-auditor-scan/report test files. No importer changes.
